+++ Runtime/ISaveGameManager.cs
@@ -15,25 +15,31 @@ namespace UniSharper.Data.SaveGame
         /// </summary>
         /// <value>The store path where data to save.</value>
         string StorePath { get; }
-        
+
         /// <summary>
-        /// Gets the encryption provider of save game data.
+        /// Gets the crypto provider.
         /// </summary>
         /// <value>The encryption provider.</value>
-        ISaveGameDataCryptoProvider SaveGameDataCryptoProvider { get; }
-        
+        ICryptoProvider CryptoProvider { get; }
+
+        /// <summary>
+        /// Gets the compression provider.
+        /// </summary>
+        ICompressionProvider CompressionProvider { get; }
+
         /// <summary>
         /// Initializes save game manager.
         /// </summary>
-        /// <param name="storePath"></param>
-        /// <param name="saveGameDataCryptoProvider"></param>
-        void Initialize(string storePath = null, ISaveGameDataCryptoProvider saveGameDataCryptoProvider = null);
+        /// <param name="storePath">The store path where data to save.</param>
+        /// <param name="cryptoProvider">The crypto provider.</param>
+        /// <param name="compressionProvider">The compression provider.</param>
+        void Initialize(string storePath = null, ICryptoProvider cryptoProvider = null, ICompressionProvider compressionProvider = null);
 
         /// <summary>
         /// Get the file path to store with specified file name.
         /// </summary>
         /// <param name="name">The specified file name. </param>
-        /// <param name="autoCreateFolder">Whether create folder if the folder under path do not exists. </param>
+        /// <param name="autoCreateFolder">Whether create folder if the folder under path do not exist. </param>
         /// <returns></returns>
         string GetFilePath(string name, bool autoCreateFolder = false);
 
@@ -54,7 +60,7 @@ namespace UniSharper.Data.SaveGame
         /// <param name="name">The name of save data.</param>
         /// <param name="data">
         /// When this method returns, contains the save data of <see cref="System.String"/>, if the
-        /// operation succeeded, or an null value if the operation failed.
+        /// operation succeeded, or a null value if the operation failed.
         /// <c>true</c> if the game data of <see cref="System.String"/> is loaded, <c>false</c> otherwise.
         /// </returns>
         bool TryLoadGame(string name, out string data);
@@ -65,7 +71,7 @@ namespace UniSharper.Data.SaveGame
         /// <param name="name">The name of save data.</param>
         /// <param name="data">
         /// When this method returns, contains the raw save data of game, if the operation
-        /// succeeded, or an null value if the operation failed.
+        /// succeeded, or a null value if the operation failed.
         /// </param>
         /// <returns><c>true</c> if the raw data of game is loaded, <c>false</c> otherwise.</returns>
         bool TryLoadGameData(string name, out byte[] data);
@@ -89,10 +95,13 @@ namespace UniSharper.Data.SaveGame
         /// </summary>
         /// <param name="name">The name of save data.</param>
         /// <param name="data">The save data of <see cref="System.String"/>.</param>
-        /// <param name="encrypt">
-        /// if set to <c>true</c> [encrypt the game data of <see cref="System.String"/> before saving].
-        /// </param>
-        void SaveGame(string name, string data, bool encrypt = true);
+        /// <param name="encrypt">if set to <c>true</c> [encrypt the game data of <see cref="System.String"/> before saving].</param>
+        /// <param name="compress">if set to <c>true</c> [compress the game data of <see cref="System.String"/> before saving].</param>
+        /// <returns><c>true</c> if data successfully saved, <c>false</c> otherwise.</returns>
+        bool SaveGame(string name,
+            string data,
+            bool encrypt = true,
+            bool compress = false);
 
         /// <summary>
         /// Saves the raw data of game.
@@ -100,8 +109,12 @@ namespace UniSharper.Data.SaveGame
         /// <param name="name">The name of save data.</param>
         /// <param name="data">The raw save data of game.</param>
         /// <param name="encrypt">if set to <c>true</c> [encrypt the raw data of game before saving].</param>
-        /// <exception cref="ArgumentNullException">name or rawData</exception>
-        void SaveGameData(string name, byte[] data, bool encrypt = true);
+        /// <param name="compress">if set to <c>true</c> [compress the raw data of game before saving].</param>
+        /// <returns><c>true</c> if data successfully saved, <c>false</c> otherwise.</returns>
+        bool SaveGameData(string name,
+            byte[] data,
+            bool encrypt = true,
+            bool compress = false);
 
         /// <summary>
         /// Deletes the save data.